LSOA-109 Filter observations to delete by owner

+++ kidviz/views.py
@@ -1016,6 +1016,6 @@ class WorkQueue(LoginRequiredMixin, ListView):
 
 class RemoveDraft(LoginRequiredMixin, View):
     def post(self, request, pk):
-        Observation.objects.filter(pk=pk, is_draft=True).delete()
+        Observation.objects.filter(pk=pk, is_draft=True, owner=request.user).delete()
         messages.add_message(self.request, messages.SUCCESS, 'Draft removed.')
         return HttpResponseRedirect(reverse_lazy('work-queue'))
